feat: move spellcheck pref into defaults preferences section
